requirements.txt
beautifulsoup4==4.12.3
langchain==0.2.17
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.38.0
flask~=3.0.3
pinecone~=5.3.1
langchain-community==0.2.7
langchain-core~=0.2.43
gunicorn